refactor: 카프카 폴더 구성
