+++ backend/src/main/java/com/Catch_Course/domain/reservation/service/ReservationService.java
@@ -4,8 +4,8 @@ import com.Catch_Course.domain.course.entity.Course;
 import com.Catch_Course.domain.course.repository.CourseRepository;
 import com.Catch_Course.domain.member.entity.Member;
 import com.Catch_Course.domain.member.repository.MemberRepository;
-import com.Catch_Course.domain.reservation.dto.ReservationProducer;
-import com.Catch_Course.domain.reservation.dto.ReservationRequest;
+import com.Catch_Course.global.kafka.producer.ReservationProducer;
+import com.Catch_Course.global.kafka.dto.ReservationRequest;
 import com.Catch_Course.domain.reservation.entity.Reservation;
 import com.Catch_Course.domain.reservation.entity.ReservationStatus;
 import com.Catch_Course.domain.reservation.repository.ReservationRepository;
