+++ backend/src/main/java/com/Catch_Course/global/kafka/producer/ReservationProducer.java
@@ -1,5 +1,6 @@
-package com.Catch_Course.domain.reservation.dto;
+package com.Catch_Course.global.kafka.producer;
 
+import com.Catch_Course.global.kafka.dto.ReservationRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
